# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Jimu 后端框架 API - 提供用户认证、权限管理、角色管理、系统监控等功能",
        "title": "Jimu API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/oauth/{provider}/callback": {
            "get": {
                "description": "处理第三方授权回调，签发 JWT",
                "tags": [
                    "OAuth"
                ],
                "summary": "OAuth 回调",
                "parameters": [
                    {
                        "type": "string",
                        "description": "提供商",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "授权码",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "状态",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    }
                }
            }
        },
        "/oauth/{provider}/login": {
            "get": {
                "description": "重定向到第三方授权页",
                "tags": [
                    "OAuth"
                ],
                "summary": "OAuth 登录跳转",
                "parameters": [
                    {
                        "type": "string",
                        "description": "提供商名（内置 google/github/wechat，或配置了 issuer_url 的自定义 OIDC 提供商）",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "details": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}